requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiolimiter==1.1.0
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==3.7.1
anytree==2.12.1
APScheduler==3.10.4
asgiref==3.8.1
asttokens==2.4.1
asyncer==0.0.7
attrs==24.2.0
autograd==1.7.0
azure-common==1.1.28
azure-core==1.32.0
azure-identity==1.19.0
azure-search-documents==11.5.2
azure-storage-blob==12.23.1
backoff==2.2.1
bcrypt==4.2.0
beartype==0.18.5
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
chevron==0.14.0
chroma-hnswlib==0.7.3
chromadb==0.5.3
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.0
cramjam==2.9.0
cryptography==42.0.8
cycler==0.12.1
dask==2024.10.0
dask-expr==1.1.16
dataclasses-json==0.6.7
datashaper==0.0.49
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
devtools==0.12.2
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
durationpy==0.9
email_validator==2.2.0
environs==11.0.0
executing==2.1.0
fastapi==0.110.3
fastapi-cli==0.0.5
fastapi-sso==0.10.0
fastparquet==2024.5.0
filelock==3.16.1
filetype==1.2.0
FLAML==2.3.2
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.10.0
ftfy==6.3.1
future==1.0.0
gensim==4.3.3
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.35.0
googleapis-common-protos==1.65.0
graphrag==0.3.6
graspologic==3.4.1
graspologic-native==1.2.1
grpcio==1.67.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.2
humanfriendly==10.0
hyppo==0.4.0
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.4.5
ipython==8.29.0
isodate==0.7.2
jedi==0.19.1
Jinja2==3.1.4
jiter==0.7.0
joblib==1.4.2
json_repair==0.28.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
kubernetes==31.0.0
lancedb==0.13.0
latexcodec==3.0.0
Lazify==0.4.0
lightgbm==4.5.0
linkify-it-py==2.0.3
litellm==1.51.3
literalai==0.0.623
llvmlite==0.43.0
locket==1.0.0
markdown-it-py==3.0.0
markdownify==0.13.1
marker-pdf==0.3.10
MarkupSafe==3.0.2
marshmallow==3.23.1
matminer==0.9.3
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
monty==2024.10.21
mpmath==1.3.0
msal==1.31.0
msal-extensions==1.2.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.15.2
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
ollama==0.3.3
onnxruntime==1.20.0
openai==1.53.0
opencv-python==4.10.0.84
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.28.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.28.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.11
overrides==7.7.0
packaging==23.2
palettable==3.3.3
pandas==2.2.3
parso==0.8.4
partd==1.4.2
patsy==0.5.6
pdftext==0.3.18
pillow==10.4.0
plotly==5.24.1
portalocker==2.10.1
posthog==3.7.0
POT==0.9.4
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==4.25.3
psutil==6.1.0
pure_eval==0.2.3
py==1.11.0
pyaml-env==1.2.1
pyarrow==15.0.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyautogen==0.2.32
pybtex==0.24.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.9.0
pylance==0.17.0
pymatgen==2024.10.29
pymongo==4.10.1
PyNaCl==1.5.0
pynndescent==0.5.13
pyparsing==3.2.0
pypdf==5.1.0
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-catmap==0.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-multipart==0.0.9
python-socketio==5.11.4
pytz==2024.2
pywin32==308
PyYAML==6.0.2
RapidFuzz==3.10.1
ratelimiter==1.2.0.post0
redis==5.2.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
retry==0.9.2
rich==13.9.4
rpds-py==0.20.1
rq==2.0.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
sentence-transformers==3.2.1
setuptools==75.1.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
spglib==2.5.0
stack-data==0.6.3
starlette==0.37.2
statsmodels==0.14.4
streamlit==1.40.2
surya-ocr==0.6.13
swifter==1.4.0
sympy==1.13.1
syncer==2.0.3
tabled-pdf==0.1.4
tabulate==0.9.0
tenacity==9.0.0
termcolor==2.5.0
texify==0.2.1
textual==0.74.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.20.1
toml==0.10.2
tomli==2.0.2
toolz==1.0.0
torch==2.5.1
tornado==6.4.2
tqdm==4.66.6
traitlets==5.14.3
transformers==4.46.1
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uc-micro-py==1.0.3
umap-learn==0.5.7
uncertainties==3.2.2
uptrace==1.27.0
urllib3==2.2.3
uvicorn==0.25.0
watchdog==6.0.0
watchfiles==0.20.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.1
wheel==0.44.0
wrapt==1.16.0
wsproto==1.2.0
xgboost==2.1.3
yarl==1.17.1
zipp==3.20.2
